Adds ability to run a process
`EventEmitter` module propagates messages asynchronously from
`Producer`s to `Channels`.  `Channel` provides a `#watch` method which
blocks, watching for messages.  The `#finalize` method must be called on
it to avoid a deadlock when no more messages are to be sent.

`Process` spawns a process, and utilizes `OutputHandler`s to emit events
on output.  Also emits an event when the process exits.

`Executor` creates a channel, binds the given `Process` (or in the
future another kind of runnable) to it, then dispatches output events to
the `Loggers::Console` if desired according to its given config.
`Executor` closes its channel when.

This architecture is very thread-heavy.  Async IO would probably be
faster and less complicated, but Ruby's stdlib doesn't seem to have good
primitives for that and C extensions would be a bother for scripts which
vendor proclib.